+++ src/main/java/net/mcreator/ui/component/TechnicalButton.java
@@ -19,12 +19,14 @@
 
 package net.mcreator.ui.component;
 
+import net.mcreator.ui.component.entries.JEntriesList;
+
 import javax.swing.*;
 
 /**
  * This is a special subtype of regular buttons instances of which used as components of
  * {@link net.mcreator.ui.modgui.ModElementGUI ModElementGUIs} perform technical operations when pressed (for instance,
- * add fresh entries to a {@link net.mcreator.ui.minecraft.JEntriesList JEntriesList} or import missing resources)
+ * add fresh entries to a {@link JEntriesList JEntriesList} or import missing resources)
  * and should not trigger {@link net.mcreator.ui.modgui.ModElementChangedListener ModElementChangedListeners}.
  */
 public class TechnicalButton extends JButton {
