+++ src/main/java/net/mcreator/ui/minecraft/boundingboxes/JBoundingBoxList.java
@@ -28,7 +28,7 @@ import net.mcreator.ui.MCreator;
 import net.mcreator.ui.component.TechnicalButton;
 import net.mcreator.ui.help.IHelpContext;
 import net.mcreator.ui.init.L10N;
-import net.mcreator.ui.minecraft.JSimpleEntriesList;
+import net.mcreator.ui.component.entries.JSingleEntriesList;
 import net.mcreator.workspace.resources.Model;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -41,7 +41,7 @@ import java.util.List;
 import java.util.Objects;
 import java.util.function.Supplier;
 
-public class JBoundingBoxList extends JSimpleEntriesList<JBoundingBoxEntry, IBlockWithBoundingBox.BoxEntry> {
+public class JBoundingBoxList extends JSingleEntriesList<JBoundingBoxEntry, IBlockWithBoundingBox.BoxEntry> {
 
 	private static final Logger LOG = LogManager.getLogger(JBoundingBoxList.class);
