+++ src/main/java/net/mcreator/ui/minecraft/spawntypes/JSpawnEntriesList.java
@@ -22,7 +22,7 @@ import net.mcreator.element.types.Biome;
 import net.mcreator.ui.MCreator;
 import net.mcreator.ui.help.IHelpContext;
 import net.mcreator.ui.init.L10N;
-import net.mcreator.ui.minecraft.JSimpleEntriesList;
+import net.mcreator.ui.component.entries.JSingleEntriesList;
 
 import javax.swing.*;
 import java.awt.*;
@@ -30,7 +30,7 @@ import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
 
-public class JSpawnEntriesList extends JSimpleEntriesList<JSpawnListEntry, Biome.SpawnEntry> {
+public class JSpawnEntriesList extends JSingleEntriesList<JSpawnListEntry, Biome.SpawnEntry> {
 
 	public JSpawnEntriesList(MCreator mcreator, IHelpContext gui) {
 		super(mcreator, gui);
